pyproject.toml
[project]
name = "group-sense"
dynamic = ["version"]
description = "Group Sense"
authors = [
    { name = "Martin Krasser", email = "martin@gradion.ai" },
]
requires-python = ">=3.11,<3.15"
readme = "README.md"
license = "Apache-2.0"
dependencies = [
    "pydantic-ai>=1.3.0",
]

[tool.uv]
default-groups = [
    "docs",
    "dev",
]

[dependency-groups]
docs = [
    "griffe-pydantic>=1.1.8",
    "mkdocs>=1.6.1,<2",
    "mkdocs-material>=9.5.48,<10",
    "mkdocstrings-python>=1.12.2,<2",
    "mkdocs-llmstxt>=0.4.0",
    "click<=8.2.1",
]
dev = [
    "invoke>=2.2,<3",
    "pre-commit>=4.3.0,<5",
    "pytest>=8.4.2,<9",
    "pytest-asyncio>=1.2.0,<2",
    "pytest-cov>=4.1.0,<5",
    "types-aiofiles>=25.1.0.20251011,<26",
    "group-terminal<=0.0.2",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
pattern = "default-unprefixed"
style = "pep440"

[tool.hatch.build.targets.sdist]
include = ["group_sense"]

[tool.hatch.build.targets.wheel]
include = ["group_sense"]

[tool.ruff]
line-length = 120

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "aiofiles.*"
]
ignore_missing_imports = true
